# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards Management API
  description: API for managing bank cards with admin functionality
  contact:
    name: API Support
    email: api@example.com
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication Controller
      summary: Register new user
      description: Register a new user with username and password
      operationId: register
      requestBody:
        description: Authentication credentials for registration
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
            example:
              username: new_user
              password: securePassword123
        required: true
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        "400":
          description: Validation error or invalid request
          content:
            application/json:
              examples:
                Validation error:
                  description: Validation error
                  value:
                    error: password - Password should be specified;
                    timestamp: "1757618375981"
                User exists:
                  description: User exists
                  value:
                    error: username - This username is already in use;
                    timestamp: "1757618375981"
        "409":
          description: Username already exists
          content:
            application/json:
              example:
                error: Such username already exists
                timestamp: "1757618375981"
  /auth/refresh-token:
    post:
      tags:
        - Authentication Controller
      summary: Refresh JWT tokens
      description: Obtain new access and refresh tokens using a valid refresh token
      operationId: refreshToken
      requestBody:
        description: Refresh token request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        required: true
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWTResponse"
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          description: Invalid refresh token format
          content:
            application/json:
              example:
                error: Invalid refresh token
                timestamp: "1757618375981"
        "404":
          description: User not found
          content:
            application/json:
              example:
                error: User not found
                timestamp: "1757618375981"
  /auth/login:
    post:
      tags:
        - Authentication Controller
      summary: Login to the system
      description: Authenticate user with username and password to receive JWT tokens
      operationId: login
      requestBody:
        description: Authentication credentials for login
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
            example:
              username: existing_user
              password: userPassword123
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWTResponse"
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          description: Validation error
          content:
            application/json:
              examples:
                Validation error:
                  description: Validation error
                  value:
                    error: password - Password should be specified;
                    timestamp: "1757618375981"
        "404":
          description: User not found
          content:
            application/json:
              example:
                error: User not found
                timestamp: "1757618375981"
        "401":
          description: Invalid credentials
          content:
            application/json:
              example:
                error: Wrong credentials
                timestamp: "1757618375981"
...
components:
  schemas:
    AuthRequest:
      type: object
      description: Authentication request DTO containing login/register credentials
      properties:
        username:
          type: string
          description: Username for authentication
          example: john_doe
          maxLength: 64
          minLength: 5
        password:
          type: string
          description: Password for authentication
          example: securePassword123
          maxLength: 60
          minLength: 5
      required:
        - password
        - username
    RefreshTokenRequest:
      type: object
      description: Refresh token request DTO
      properties:
        refreshToken:
          type: string
          description: Refresh token to obtain new access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFt...
          maxLength: 255
          minLength: 0
      required:
        - refreshToken
    JWTResponse:
      type: object
      description: JWT tokens response DTO
      properties:
        accessToken:
          type: string
          description: JWT access token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJCYW5rIGRldGFpbHMi...
        refreshToken:
          type: string
          description: JWT refresh token for obtaining new access tokens
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJCYW5rIGRldGFpbHMi...
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT